requirements.txt
Flask==2.1.2
aws-cdk.core==2.0.0
aws-cdk.aws-s3==2.0.0
aws-cdk.aws-ec2==2.0.0
aws-cdk.aws-lambda==2.0.0
aws-cdk.aws-apigateway==2.0.0
aws-cdk.aws-s3-deployment==2.0.0
